task-1_Unit-Converter/pyproject.toml
[project]
name = "task-1-unit-converter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.4",
    "streamlit>=1.42.2",
]
